component: docker_build
